Move more scene tests to oak
